requirements.txt
appdirs==1.4.4
beautifulsoup4==4.12.2
blinker==1.6.3
bs4==0.0.1
certifi==2023.7.22
charset-normalizer==3.3.1
click==8.1.7
Flask==3.0.0
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.2
pyee==8.2.2
pyppeteer==1.0.2
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.31.0
six==1.16.0
soupsieve==2.5
tqdm==4.66.1
urllib3==1.26.18
webdriver-manager==4.0.1
websockets==10.4
Werkzeug==3.0.0
zipp==3.17.0
